+++ trash/testAleatorio.c
@@ -23,22 +23,22 @@ double cpu_time_used;
 #define pbe 0.33 //probabilidad de busqueda exitosa
 #define pbi 0.17 //probabilidad de busqueda infructuosa
 
-long cant_in=(long)(n*pi); //cantidad de inserciones
-long cant_be=(long)(n*pbe); //cantidad de busquedas exitosas
-long cant_bi=(long)(n*pbi); //cantidad de bisquedas infructuosas
+long long cant_in=(long long)(n*pi); //cantidad de inserciones
+long long cant_be=(long long)(n*pbe); //cantidad de busquedas exitosas
+long long cant_bi=(long long)(n*pbi); //cantidad de bisquedas infructuosas
 
 int secuencia[n]; //arreglo de largo 100 que cant_nodosendra las operaciones
 
 //busca un elemento en un arreglo, si lo encuentra retorna 1 sino entrega 0
-int search_in_array(long arreglo[], long busqueda, long ene) {
+int search_in_array(long long arreglo[], long long busqueda, long long ene) {
     for (int i = 0; i < ene; i++) {
         if (arreglo[i] == busqueda) return 1;
     }
     return 0;
 }
 
-long get_random_not_repeted(long arreglo[],long size,long max_numero){
-    long random;
+long long get_random_not_repeted(long long arreglo[],long long size,long long max_numero){
+    long long random;
     do{
         random = (rand() % max_numero);
     } while (search_in_array(arreglo,random,size));
@@ -80,21 +80,21 @@ void aleatorio(double times[]){
     NodeBt_16 *rootBt_16 = NULL;
     NodeBt_256 *rootBt_256 = NULL;
     NodeBt_4096 *rootBt_4096 = NULL;
-    long *nodos= (long *)malloc(sizeof(long)*(n/2));
-    long *params = (long *)malloc(sizeof(long)*n);
-    long param;
-    long index = 0; //indice de operacion
-    long cant_nodos=0;
-    long max_numero = pow(2,32); //maximo numero posible
-    long random = 0; //numero obtenido al azar
+    long long *nodos= (long long *)malloc(sizeof(long long)*(n/2));
+    long long *params = (long long *)malloc(sizeof(long long)*n);
+    long long param;
+    long long index = 0; //indice de operacion
+    long long cant_nodos=0;
+    long long max_numero = pow(2,32); //maximo numero posible
+    long long random = 0; //numero obtenido al azar
     int operacion; //operacion obtenida de la secuencia
 
     printf("Loaging\n");
 
-    for (long i=0; i<n; i++) {
+    for (long long i=0; i<n; i++) {
         operacion = secuencia[i];
         if (operacion==0){
-            long random = get_random_not_repeted(nodos,cant_nodos,max_numero);
+            long long random = get_random_not_repeted(nodos,cant_nodos,max_numero);
             nodos[cant_nodos]=random; 
             params[index]=random; 
             cant_nodos++;
@@ -102,14 +102,14 @@ void aleatorio(double times[]){
             printf("[%lld,%lld]\n",i+1,n);
         }
         else if (operacion==1){
-            long indice = (rand() % cant_nodos);
+            long long indice = (rand() % cant_nodos);
             random=nodos[indice];
             params[index]=random;
             index++;
             printf("[%lld,%lld]\n",i+1,n);
         }
         else{
-            long random = get_random_not_repeted(nodos,cant_nodos,max_numero);
+            long long random = get_random_not_repeted(nodos,cant_nodos,max_numero);
             params[index]=random; 
             index++;
             printf("[%lld,%lld]\n",i+1,n);
@@ -119,7 +119,7 @@ void aleatorio(double times[]){
     //printf(" ACA PARTE EL ABB\n");
     
     start = clock(); //inicio el temporizador
-    for (long i=0; i < n; i++){ //recorro mi secuencia
+    for (long long i=0; i < n; i++){ //recorro mi secuencia
         operacion = secuencia[i]; //obtengo la operacion
         param = params[i];
         //printf("el paramentro es %d\n",param);
@@ -143,7 +143,7 @@ void aleatorio(double times[]){
     //printf(" ACA PARTE EL AVL\n");
 
     start = clock(); //inicio el temporizador
-    for (long i=0; i < n; i++){ //recorro mi secuencia
+    for (long long i=0; i < n; i++){ //recorro mi secuencia
         operacion = secuencia[i]; //obtengo la operacion
         param = params[i];
         //printf("el paramentro es %d\n",param);
@@ -167,7 +167,7 @@ void aleatorio(double times[]){
    // printf("ACA PARTE EL SPLAY");
 
     start = clock(); //inicio el temporizador
-    for (long i=0; i < n; i++){ //recorro mi secuencia
+    for (long long i=0; i < n; i++){ //recorro mi secuencia
         operacion = secuencia[i]; //obtengo la operacion
         param = params[i];
         //printf("el paramentro es %d\n",param);
@@ -191,7 +191,7 @@ void aleatorio(double times[]){
     //printf("ACA PARTE EL ARBOL B 16\n");
 
     start = clock(); //inicio el temporizador
-    for (long i=0; i < n; i++){ //recorro mi secuencia
+    for (long long i=0; i < n; i++){ //recorro mi secuencia
         operacion = secuencia[i]; //obtengo la operacion
         param = params[i];
         //printf("el paramentro es %d\n",param);
@@ -215,7 +215,7 @@ void aleatorio(double times[]){
     //printf("ACA PARTE EL ARBOL B 256\n");
 
     start = clock(); //inicio el temporizador
-    for (long i=0; i < n; i++){ //recorro mi secuencia
+    for (long long i=0; i < n; i++){ //recorro mi secuencia
         operacion = secuencia[i]; //obtengo la operacion
         param = params[i];
         //printf("el paramentro es %d\n",param);
@@ -239,7 +239,7 @@ void aleatorio(double times[]){
     //printf("ACA PARTE EL ARBOL B 4096\n");
 
     start = clock(); //inicio el temporizador
-    for (long i=0; i < n; i++){ //recorro mi secuencia
+    for (long long i=0; i < n; i++){ //recorro mi secuencia
         operacion = secuencia[i]; //obtengo la operacion
         param = params[i];
         //printf("el paramentro es %d\n",param);
